[CSL-1926] Create Cardano.Wallet.API.Request.Pagination module
